Init : 협업을 위해 Class 단위로 스크립트 생성
협업을 위해 Class 단위로 스크립트 생성
